requirement.txt
cachetools==5.1.0
certifi==2022.5.18
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Flask==2.1.2
Flask-MySQLdb==1.0.1
google-api-core==2.8.0
google-api-python-client==2.48.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-vision==2.7.2
googleapis-common-protos==1.56.1
greenlet==1.1.2
grpcio==1.46.1
grpcio-status==1.46.1
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.0
numpy==1.22.3
pandas==1.4.2
proto-plus==1.20.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.36
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
